Fix issues with uncast enum values
- Dont use enum for default uint64 initializier value
- Cast enums to integers for gmock matching

+++ src/ApplicationSampler.cpp
@@ -120,7 +120,7 @@ namespace geopm
         , m_is_cpu_active(is_cpu_active)
         , m_update_time({{0, 0}})
         , m_is_first_update(true)
-        , m_hint_last(m_num_cpu, GEOPM_REGION_HINT_UNSET)
+        , m_hint_last(m_num_cpu, uint64_t(GEOPM_REGION_HINT_UNSET))
     {
         if (m_is_cpu_active.empty()) {
             m_is_cpu_active.resize(m_num_cpu, false);
